Update the files

+++ parts/linux/cloud-init/artifacts/aks-local-dns.sh
@@ -29,7 +29,7 @@ LOCAL_DNS_CORE_FILE_PATH="/opt/azure/aks-local-dns/Corefile"
 
 if [ "${UPSTREAM_DNS_SERVERS_FROM_VNET}" != "${DEFAULT_UPSTREAM_DNS_SERVER_IP}" ]; then
     sed -ie "s/${DEFAULT_UPSTREAM_DNS_SERVER_IP}/${UPSTREAM_DNS_SERVERS_FROM_VNET}/g" "${LOCAL_DNS_CORE_FILE_PATH}"
-    printf "Replaced '${DEFAULT_UPSTREAM_DNS_SERVER_IP}' with '${UPSTREAM_DNS_SERVERS_FROM_VNET}' in '${LOCAL_DNS_CORE_FILE_PATH}' \n"
+    printf "Updated DNSIP from '%s' to '%s' in '%s'\n" "${DEFAULT_UPSTREAM_DNS_SERVER_IP}" "${UPSTREAM_DNS_SERVERS_FROM_VNET}" "${LOCAL_DNS_CORE_FILE_PATH}"
 fi
 
 
@@ -38,14 +38,15 @@ fi
 # These rules skip conntrack for DNS traffic to save conntrack table
 # space. OUTPUT rules affect node services and hostNetwork: true pods.
 # PREROUTING rules affect traffic from regular pods.
-IPTABLES='iptables -w -t raw -m comment --comment "AKS Local DNS: skip conntrack"'
+IPTABLES='iptables -w -t raw -m comment --comment "aks-local-dns: skip conntrack"'
 IPTABLES_RULES=()
 for CHAIN in OUTPUT PREROUTING; do
-for IP in ${NODE_LISTENER_IP} ${CLUSTER_LISTENER_IP}; do
-for PROTO in tcp udp; do
-    IPTABLES_RULES+=("${CHAIN} -p ${PROTO} -d ${IP} --dport 53 -j NOTRACK")
-done; done; done
-
+   for IP in ${NODE_LISTENER_IP} ${CLUSTER_LISTENER_IP}; do
+       for PROTO in tcp udp; do
+           IPTABLES_RULES+=("${CHAIN} -p ${PROTO} -d ${IP} --dport 53 -j NOTRACK")
+       done
+    done
+done
 
 # cleanup function: will be run on script exit/crash to revert config
 # -------------------------------------------------------------------------------------------
@@ -62,17 +63,16 @@ function cleanup {
 
     # Revert the changes made to the DNS configuration if present
     if [ -f ${NETWORK_DROPIN_FILE} ]; then
-        printf "reverting dns configuration by removing ${NETWORK_DROPIN_FILE}\n"
+        printf "Reverting DNS configuration by removing %s\n" "${NETWORK_DROPIN_FILE}"
         /bin/rm -f ${NETWORK_DROPIN_FILE}
         networkctl reload
     fi
 
-    # Trigger coredns shutdown, if runnin
+    # Trigger coredns shutdown, if running
     if [ ! -z "${COREDNS_PID:-}" ]; then
         if ps ${COREDNS_PID} >/dev/null; then
             if [[ ${COREDNS_SHUTDOWN_DELAY} -gt 0 ]]; then
-                # Wait after removing iptables rules and DNS configuration so that we can let connections
-                # transition.
+                # Wait after removing iptables rules and DNS configuration so that we can let connections transition.
                 printf "sleeping ${COREDNS_SHUTDOWN_DELAY} seconds to allow connections to terminate\n"
                 sleep ${COREDNS_SHUTDOWN_DELAY}
             fi
@@ -106,7 +106,7 @@ fi
 # coredns: extract from image
 # -------------------------------------------------------------------------------------------
 if [ ! -x ${SCRIPT_PATH}/coredns ]; then
-    printf "extracting coredns from docker image: ${COREDNS_IMAGE}\n"
+    printf "extracting coredns from docker image: %s\n" "${COREDNS_IMAGE}"
     CTR_TEMP="$(mktemp -d)"
 
     # Set a trap to clean up the temp directory if anything fails
@@ -114,23 +114,36 @@ if [ ! -x ${SCRIPT_PATH}/coredns ]; then
         # Disable error handling so that we don't get into a recursive loop
         set +e
         printf 'Error extracting coredns\n'
-        ctr -n k8s.io images unmount ${CTR_TEMP}
-        rm -rf ${CTR_TEMP}
+        ctr -n k8s.io images unmount "${CTR_TEMP}" >/dev/null
+        rm -rf "${CTR_TEMP}"
     }
     trap cleanup_coredns_import EXIT ABRT ERR INT PIPE QUIT TERM
 
+    # Check if the image exists in the local cache
+    if ! ctr -n k8s.io images ls | grep -q "${COREDNS_IMAGE}"; then
+        printf "CoreDNS image not found locally. Attempting to pull from mcr.microsoft.com...\n"
+        if ! ctr -n k8s.io images pull "${COREDNS_IMAGE}"; then
+            printf 'Error: failed to pull coredns image: %s\n' "${COREDNS_IMAGE}"
+            exit 1
+        fi
+    fi
+
     # Mount the coredns image to the temporary directory
-    ctr -n k8s.io images mount ${COREDNS_IMAGE} ${CTR_TEMP} >/dev/null
+    if ctr -n k8s.io images mount "${COREDNS_IMAGE}" "${CTR_TEMP}" >/dev/null; then
 
-    # Copy coredns to SCRIPT_PATH
-    cp ${CTR_TEMP}/coredns ${SCRIPT_PATH}/coredns
+      # Copy coredns to SCRIPT_PATH
+      cp "${CTR_TEMP}/coredns" "${SCRIPT_PATH}/coredns"
 
-    # Umount and clean up the temporary directory
-    ctr -n k8s.io images unmount ${CTR_TEMP} >/dev/null
-    rm -rf "${CTR_TEMP}"
+      # Umount and clean up the temporary directory
+      ctr -n k8s.io images unmount "${CTR_TEMP}" >/dev/null
+      rm -rf "${CTR_TEMP}"
 
-    # Clear the trap
-    trap - EXIT ABRT ERR INT PIPE QUIT TERM
+      # Clear the trap
+      trap - EXIT ABRT ERR INT PIPE QUIT TERM
+    else
+        printf 'Error: failed to mount coredns image: %s\n' "${COREDNS_IMAGE}"
+        exit 1
+    fi
 fi
 
 # Enable the cleanup function now that we have a coredns binary
